agregando la mesa ya funcionando bien y agregando esta funcion a una carpeta resolvers por fuera del codigo
