+++ api/src/models/Mesas.js
@@ -4,7 +4,6 @@ const MesasSchema = new Schema({
     users:  [{ 
         type: Schema.Types.ObjectId, ref: 'User'
     }],
-    
     linkMeet: {
         type: String
     },
